now instrumenting from AST alone
This does not yet work in all cases. Particularly, some wierd function
will no longer compile if they end in a an infinite loop that returns
inside the loop.

func() int {
	c := asdf
	for {
		if c {
			return 1
		}
		c = asdf
	}
}

It will add instrumentation like

func() int {
	// enter
	c := asdf
	for {
		// enter
		if c {
			// enter
			// exit
			return 1
		}
		c = asdf
		// exit
	}
	// exit
}

However, the ending exit insturmentation will cause a compile error
saying the function should end in a return stmt. WTF. How did that
function compile in the first place but not be able to prove that the
final exit is dead code?

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>